+++ examples/common/Wire.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "IdfArduinoCompat.h"